+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.1
+
+- Make replaceWithEmojis() a top-level function. Do away with EmojiUtil class.
+- Update homepage in pubspec.
+
 ## 1.0
 
 - Initial version, with support for GitHub Emojis.